requirements.txt
opencv-python
pyyaml
fastapi
uvicorn[standard]
python-dotenv
